# Conflicts:
#	mercury/pom.xml
#	mercury/src/main/db/mercurydb.changelog-1.2.xml
#	mercury/src/main/java/org/broadinstitute/gpinformatics/infrastructure/parsers/csv/CsvParser.java
#	mercury/src/main/java/org/broadinstitute/gpinformatics/infrastructure/search/LabVesselSearchDefinition.java
#	mercury/src/main/java/org/broadinstitute/gpinformatics/mercury/boundary/run/FingerprintResource.java
#	mercury/src/main/java/org/broadinstitute/gpinformatics/mercury/boundary/zims/IlluminaRunResource.java
#	mercury/src/main/java/org/broadinstitute/gpinformatics/mercury/control/run/ConcordanceCalculator.java
#	mercury/src/main/java/org/broadinstitute/gpinformatics/mercury/entity/run/IlluminaFlowcell.java
#	mercury/src/main/java/org/broadinstitute/gpinformatics/mercury/entity/vessel/LabVessel.java
#	mercury/src/test/java/org/broadinstitute/gpinformatics/mercury/boundary/run/SolexaRunResourceNonRestTest.java
#	mercury/src/test/java/org/broadinstitute/gpinformatics/mercury/entity/sample/ControlFixupTest.java

+++ mercury/src/main/java/org/broadinstitute/gpinformatics/athena/entity/orders/PriceAdjustment.java
@@ -15,7 +15,7 @@ public abstract class PriceAdjustment {
 
     public abstract String getCustomProductName();
 
-    public abstract Integer getAdjustmentQuantity();
+    public abstract BigDecimal getAdjustmentQuantity();
 
     public boolean hasPriceAdjustment() {
         return getAdjustmentValue() != null || StringUtils.isNotBlank(getCustomProductName()) || getAdjustmentQuantity() != null;
